made m_size of type (size_t) instead of just (int)
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@1873 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Manips/vertex_program.h
@@ -98,7 +98,7 @@ class GEM_EXTERN vertex_program : public GemBase
   GLuint 	m_programTarget;
   GLuint    m_programID;
   char*		m_programString;
-  int		m_size;
+  size_t	m_size;
   char		m_buf[MAXPDSTRING];
   
   float		m_param[4];
